Don't hold the GIL during calls
Release the GIL during calls to the C library. This allows other threads in the
Python process to progress while the library is busy talking etc.

Since the Festival library is not guaranteed to be thread-safe, we now use a
mutex to guard calls into it, thus ensuring that only one thrad at a time may
use the library. Previously the GIL was effectively doing this for us.